num_sdf to long type

+++ object_wise/dqn/replay_buffer.py
@@ -60,8 +60,8 @@ class ReplayBuffer(object):
             torch.FloatTensor(self.reward[ind]).to(self.device),
             torch.FloatTensor(self.not_done[ind]).to(self.device),
             torch.FloatTensor(self.goal[ind]).to(self.device),
-            torch.FloatTensor(self.numblocks[ind]).to(self.device),
-            torch.FloatTensor(self.next_numblocks[ind]).to(self.device)
+            torch.LongTensor(self.numblocks[ind]).to(self.device),
+            torch.LongTensor(self.next_numblocks[ind]).to(self.device),
         ]
         if self.save_img:
             data_bath.append(torch.FloatTensor(self.state_im[ind]).to(self.device))
@@ -187,8 +187,8 @@ class PER(object):
             torch.FloatTensor(self.reward[ind]).to(self.device),
             torch.FloatTensor(self.not_done[ind]).to(self.device),
             torch.FloatTensor(self.goal[ind]).to(self.device),
-            torch.FloatTensor(self.numblocks[ind]).to(self.device),
-            torch.FloatTensor(self.next_numblocks[ind]).to(self.device)
+            torch.LongTensor(self.numblocks[ind]).to(self.device),
+            torch.LongTensor(self.next_numblocks[ind]).to(self.device),
         ]
         if self.save_img:
             data_bath.append(torch.FloatTensor(self.state_im[ind]).to(self.device))
